email-settings(fix): preserve state for unknown reducer actions
